Liquidslider - v1
liquidslider is a jquery slider that allows for both page sliding and
image fading. This slider also allows you to track clicks of the
navigation as in-page-events by Google Analytics (SEO enhancement).
This slider has a 'thumbnail' mode (thumbnails are the linked to the
slide) or and overlay (small button links that overlay the slider
image). Additionally there is a consoleDebug function that I use to
output values to the chrome or firefox debug console. This slider is
not intended for novice developers